It's turtles all the way down.

+++ src/main/java/com/openlattice/ids/IdsGeneratingEntryProcessor.java
@@ -32,7 +32,7 @@ import java.util.UUID;
 /**
  * @author Matthew Tamayo-Rios &lt;matthew@openlattice.com&gt;
  */
-public class IdsGeneratingEntryProcessor extends AbstractRhizomeEntryProcessor<Integer, Range, List<UUID>>
+public class IdsGeneratingEntryProcessor extends AbstractRhizomeEntryProcessor<Long, Range, List<UUID>>
         implements Offloadable {
     private final int count;
 
@@ -41,7 +41,7 @@ public class IdsGeneratingEntryProcessor extends AbstractRhizomeEntryProcessor<I
     }
 
     @Override
-    public List<UUID> process( Entry<Integer, Range> entry ) {
+    public List<UUID> process( Entry<Long, Range> entry ) {
         final Range range = entry.getValue(); //Range should never be null in the EP.
         final UUID[] ids = getIds( range );
         entry.setValue( range );
